+++ src/blockchain/Block.java
@@ -1,3 +1,4 @@
+package blockchain;
 
 import java.nio.ByteBuffer;
 import java.security.NoSuchAlgorithmException;
