+++ src/blockchain/Hash.java
@@ -1,3 +1,4 @@
+package blockchain;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
