+++ src/driver/commands/Command.java
@@ -1,4 +1,5 @@
-
+package driver.commands;
+import driver.BlockChainDriver;
 
 public interface Command {
   public void run(BlockChainDriver instance);
