+++ src/driver/commands/MineCommand.java
@@ -1,4 +1,7 @@
-public class RemoveCommand implements Command {
+package driver.commands;
+import driver.BlockChainDriver;
+
+public class MineCommand implements Command {
 
   @Override
   public void run(BlockChainDriver instance) {
