+++ src/driver/commands/QuitCommand.java
@@ -1,8 +1,11 @@
+package driver.commands;
+import driver.BlockChainDriver;
+
 public class QuitCommand implements Command {
 
   @Override
   public void run(BlockChainDriver instance) {
-    instance.isRunning = false;
+    instance.stop();
   }
   
 }
