+++ src/driver/commands/RemoveCommand.java
@@ -1,4 +1,7 @@
-public class ReportCommand implements Command {
+package driver.commands;
+import driver.BlockChainDriver;
+
+public class RemoveCommand implements Command {
 
   @Override
   public void run(BlockChainDriver instance) {
