+++ test/BlockTest.java
@@ -2,6 +2,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import org.junit.Test;
+import blockchain.Block;
 
 public class BlockTest {
   @Test
