+++ test/HashTest.java
@@ -1,6 +1,7 @@
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import org.junit.Test;
+import blockchain.Hash;
 
 public class HashTest {
   @Test
